+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+-   `approvalConfiguration` to `Form`
+
 ## 3.0.9 (2022-08-16)
 
 ### Added