+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 -   `approvalConfiguration` to `Form`
 
+### Changed
+
+-   Numerous updates to `JWTPayload` class
+
 ## 3.0.9 (2022-08-16)
 
 ### Added