+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 -   `approvalConfiguration` to `Form`
 -   `FormApprovalCannedResponse` class
 -   `cannedResponseKey` to `FormSubmissionApproval`
+-   `secret` to `FormServerValidationConfiguration`
+-   `isValid` parameter to `FormsClient.SearchSubmissions`
 
 ### Changed