+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 -   `FormApprovalCannedResponse` class
 -   `cannedResponseKey` to `FormSubmissionApproval`
 -   `DataManagerClient` class and corresponding model classes
+-   `secret` to `FormServerValidationConfiguration`
+-   `isValid` parameter to `FormsClient.SearchSubmissions`
 
 ### Changed